ration.yaml && app.properties updated; Edible.java: signature of eat() changed; FindAppProperties.java: name changed to AppProperties.java, added new variable and initialized it in a non-static block; Animal.java: deleted a variable, added new methods, interface Edible.java implemented; Herbivorous.java && Predator.java: initialized a boolean in constructors

+++ src/model/abstraction/Edible.java
@@ -1,5 +1,5 @@
 package model.abstraction;
 
 public interface Edible {
-    void eat(int amountOfFood);
+    void eat(double amountOfFood);
 }
